Use an Int instead of a Long in parallelism combinators (#1290)
* use an Int for ZIO parallelism combinators

* use an Int for ZStream parallelism combinators

+++ test/shared/src/main/scala/zio/test/Spec.scala
@@ -66,7 +66,7 @@ final case class Spec[+L, +T](caseValue: SpecCase[L, T, Spec[L, T]]) { self =>
             ZIO.foreachPar(specs)(_.foldM(defExec)(f)).flatMap(specs => f(SuiteCase(label, specs.toVector, exec)))
           case ExecutionStrategy.ParallelN(n) =>
             ZIO
-              .foreachParN(n.toLong)(specs)(_.foldM(defExec)(f))
+              .foreachParN(n)(specs)(_.foldM(defExec)(f))
               .flatMap(specs => f(SuiteCase(label, specs.toVector, exec)))
           case ExecutionStrategy.Sequential =>
             ZIO.foreach(specs)(_.foldM(defExec)(f)).flatMap(specs => f(SuiteCase(label, specs.toVector, exec)))
